requirements.txt
Flask==2.3.2
Flask-JWT-Extended==4.5.3
bcrypt==4.0.1
psycopg2-binary==2.9.9
opencv-python==4.8.1.78
tensorflow==2.13.0
requests==2.31.0